pyproject.toml
[project]
name = "stock-arm"
version = "0.1.0"
description = "Stock Analysis & Research Manager — A股个人金融分析工具"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.19",
    "alembic>=1.13",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "apscheduler>=3.10,<4.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "litellm>=1.0",
    "cachetools>=5.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "pytest-mock>=3.12",
    "ruff>=0.3",
    "mypy>=1.8",
    "black>=24.0",
    "isort>=5.13",
    "aioresponses>=0.7",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["efinance.*", "akshare.*", "baostock.*", "tushare.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["litellm.*", "jinja2.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
